каталог  пользовательский как в админке поправить тотал для пагинации фильтр по имении добавить %цуц%
пользовательский запрос для добавления в избранные

src="http://localhost:8080/api/files?cosmeticId=2&fileName=2_1.jpg"

+++ backend/src/main/java/ru/cosmetic/server/responseDto/CosmeticsResponse.java
@@ -11,6 +11,6 @@ import java.util.List;
 @AllArgsConstructor
 @RequiredArgsConstructor
 public class CosmeticsResponse {
-    Long total;
+    Integer total;
     List<CosmeticResponse> cosmetics;
 }
